requirements.txt
faiss-cpu==1.7.4
jupyter==1.0.0
openai==0.27.7
pdf2image==1.16.3
tabulate==0.9.0
unstructured==0.7.1
tiktoken==0.4.0
python-dotenv
streamlit==1.23.1
streamlit-chat==0.0.2.2
beautifulsoup4~=4.12.2
langchain~=0.0.187
